Make it so the timeout reader is better about the first read
the http Library calls read the first time on the body with a
[]byte of size 1 just to see if anything is in the reader
this caused the timeout reader to just return the first character
and drop the rest of the filename. Not idea.
So instead anytime a read is called and we cant copy everything
to the calling byte slice we will store the leftovers and
give them to the the next read caller.