Forging a pattern for sane output styles
The process for displaying output in, at times, nested tree structure
in addition to simply ensuring that certain steps happen in the order
that a user would assume has required a less-than-minor refactor
of the processes.

Through the process of the refactor as mentioned above, many
potential failure scenarios were discovered and corrected. Most
of these corrections are accomplished with simply idempotency.

Lastly, after multiple iterations of stylistic options, it is decided
to show the full, native docker image pull process. The process
required modifications to allow printing with a prefix, so the code
to produce the output has been copied with minor alterations. The original
work has been referenced along with the accompanying license.

There is still much required refactoring before this process can be
considered complete. This is the first pass, and at this point
much, if not all, of the pattern has been forged and the required
utilities are provided.