Support mid-value equal sign in evar loading
Fixes an issue where an equal sign in the body of a multi-line environmental variable surrounded by quotes would be interpretted as the beginning of a new variable instead of just a part of the current variable. Now when a quoted variable body is found that doesn’t end in a quote, subsequent variables will be added to the body of the current variable until the closing quote is found. This prevents accidental breaking-up of multi-line variable bodies due to equal signs.